Module Profile.Switch

One of the switches above, as a value.

type profile := t

type t

val name : t -> string

The switch's function name without with_: tables, ext_keyed.

val doc : t -> string

The syntax the switch turns on, in a few words.

val is_extension : t -> bool

Whether the switch is an extension, off in djot.

val get : t -> profile -> bool

val set : t -> bool -> profile -> profile